python-rust-mongo/pyproject.toml
[tool.poetry]
name = "python-rust-mongo"
version = "0.1.0"
description = ""
authors = ["Alessandro Mancini <alessandro.mancini4@enel.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
motor = "^3.5.0"
mongo = "^0.2.0"
python-dotenv = "^1.0.1"
uvicorn = "^0.30.1"
metrics-aggregator = {path = "metrics-aggregator", develop = true}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
